pyXcomfort/mqtt/requirements.txt
paho-mqtt==1.4.0
pythoncrc==1.21
pyserial==3.4